% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Centric_Noncoding_Results_Summary_meta.R
\name{Gene_Centric_Noncoding_Results_Summary_meta}
\alias{Gene_Centric_Noncoding_Results_Summary_meta}
\title{Generates summary table and visualization for the meta-analysis of noncoding functional categories that was conducted using MetaSTAARlite}
\usage{
Gene_Centric_Noncoding_Results_Summary_meta(
  gene_centric_noncoding_jobs_num,
  input_path,
  output_path,
  gene_centric_results_name,
  ncRNA_jobs_num,
  ncRNA_input_path,
  ncRNA_output_path,
  ncRNA_results_name,
  alpha = 2.5e-06,
  alpha_ncRNA = 2.5e-06,
  ncRNA_pos = NULL,
  manhattan_plot = FALSE,
  QQ_plot = FALSE
)
}
\arguments{
\item{gene_centric_noncoding_jobs_num}{an integer which specifies the number of jobs done in the gene-centric noncoding meta-analysis.}

\item{input_path}{a character which specifies the file path to the gene-centric noncoding meta-analysis results files.}

\item{output_path}{a character which specifies the file path to the desired location of the produced summary table and visualizations for gene-centric noncoding meta-analysis.}

\item{gene_centric_results_name}{a character which specifies the name (excluding the jobs number) of the gene-centric noncoding meta-analysis results files.}

\item{ncRNA_jobs_num}{an integer which specifies the number of jobs done in the ncRNA meta-analysis.}

\item{ncRNA_input_path}{a character which specifies the file path to the ncRNA meta-analysis results files.}

\item{ncRNA_output_path}{a character which specifies the file path to the desired location of the produced summary table and visualizations for the ncRNA meta-analysis.}

\item{ncRNA_results_name}{a character which specifies the name (excluding the jobs number) of the ncRNA meta-analysis results files.}

\item{alpha}{a numeric value which specifies the desired significance threshold for the gene-centric noncoding meta-analysis (default = 2.5E-06).}

\item{alpha_ncRNA}{a numeric value which specifies the desired significance threshold for the ncRNA meta-analysis (default is 2.5E-06).}

\item{manhattan_plot}{a logical value which determines if a Manhattan plot is generated (default = FALSE).}

\item{QQ_plot}{a logical value which determines if a QQ plot is generated (default = FALSE).}
}
\description{
This function takes in objects of gene-centric noncoding meta-analysis results from MetaSTAARlite and generates a summary table,
Manhattan plot, and QQ plot for the meta-analysis of noncoding functional categories that was conducted
based on the parameters provided by the user.
}
