% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_analysis_MetaSTAARlite_cond.R
\name{individual_analysis_MetaSTAARlite_cond}
\alias{individual_analysis_MetaSTAARlite_cond}
\title{Performs conditional meta-analysis of individual variants by using the MetaSTAARlite pipeline.}
\usage{
individual_analysis_MetaSTAARlite_cond(
  individual_results,
  sample.sizes,
  sumstat.list,
  covcond.list,
  mac_cutoff = 20,
  effect.cond = "homogeneous",
  check_qc_label = FALSE
)
}
\arguments{
\item{sample.sizes}{a numeric vector with the length of \code{study.names}
indicating the sample size of each study.}

\item{sumstat.list}{a character vector containing the directories of the study-specific summary statistics file folders.}

\item{covcond.list}{a character vector containing the names of variants to be conditioned on.}

\item{mac_cutoff}{the cutoff of minimum minor allele count in
defining individual variants. Default is 20.}

\item{effect.cond}{a character vector specifying whether the effects of variants to be adjusted for
in conditional analysis are "homogeneous" or "heterogeneous". Default is "homogeneous".}

\item{check_qc_label}{a logical value indicating whether variants need to be dropped according to \code{qc_label}. Default is FALSE.
specified in \code{\link{generate_MetaSTAAR_sumstat}} and \code{\link{generate_MetaSTAAR_cov}}.
If \code{check_qc_label} is FALSE, it is assumed that no variant will be dropped (default = FALSE).}
}
\value{
a data frame containing the score test p-value and the estimated effect size of the alternative allele for each individual variant in the given genetic region.
}
\description{
This function performs meta-analysis to detect associations between a quantitative/dichotomous phenotype
and each individual variant in a genetic region by using score test.
}
