% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_analysis_MetaSTAARlite_worker.R
\name{individual_analysis_MetaSTAARlite_worker}
\alias{individual_analysis_MetaSTAARlite_worker}
\title{Performs the worker step of MetaSTAARlite for individual variants}
\usage{
individual_analysis_MetaSTAARlite_worker(
  chr,
  start_loc,
  end_loc,
  genofile,
  obj_nullmodel,
  known_loci = NULL,
  subsegment.size = 50000,
  QC_label = "annotation/filter",
  check_qc_label = FALSE,
  variant_type = c("variant", "SNV", "Indel"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = c(FALSE, TRUE),
  Annotation_name = NULL,
  silent = FALSE
)
}
\arguments{
\item{chr}{an integer which specifies the chromosome number.}

\item{start_loc}{an integer which specifies the starting location of variants to be analyzed.}

\item{end_loc}{an integer which specifies the end location of variants to be analyzed.}

\item{genofile}{an object of opened annotated GDS (aGDS) file with variant annotation information (and without genotype).}

\item{obj_nullmodel}{an object from fitting the null model, which is the
output from either \code{\link{fit_null_glm}} function for unrelated samples or
\code{\link{fit_null_glmmkin}} function for related samples in the \code{\link{STAAR}} package.}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis. Should
contain four columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT). Default is NULL.}

\item{subsegment.size}{a numeric value which specifies the size of subsegments. Default is 5e4.}

\item{QC_label}{a character specifying the channel name of the QC label in the GDS/aGDS file.
Default is "annotation/filter".}

\item{check_qc_label}{a logical value indicating whether variants need to be dropped according to \code{qc_label}
specified in \code{\link{generate_MetaSTAAR_sumstat}} and \code{\link{generate_MetaSTAAR_cov}}. Default is FALSE.}

\item{variant_type}{a character specifying the types of variants to be considered. Choices
include "SNV", "Indel", or "variant" (default = "SNV").}

\item{Annotation_dir}{a character specifying the channel name of the annotations in the aGDS file.
Default is "annotation/info/FunctionalAnnotation"}

\item{Annotation_name_catalog}{a data frame containing the annotation name and the corresponding
channel name in the aGDS file.}

\item{Use_annotation_weights}{a logical value which specifies if annotations will be used as weights
or not. Default is TRUE.}

\item{Annotation_name}{a vector of annotation names used in MetaSTAAR. Default is NULL.}

\item{silent}{a logical value which determines if the report of error messages will be suppressed. Default is FALSE.}
}
\value{
two objects. First, the data frame of all variants in the variant-set (the summary statistics file),
including the following information: chromosome (chr), position (pos), reference allele (ref),
alternative allele (alt), quality control status (qc_label, optional), alternative allele count (alt_AC), minor allele count (MAC),
minor allele frequency (MAF), study sample size (N), score statistic (U), variance (V), and
the (low-rank decomposed) dense component of the covariance file. Second, the sparse matrix of all variants in the variant-set
whose minor allele frequency is below \code{cov_maf_cutoff} (the sparse weighted
covariance file), stored as a rectangle format.
}
\description{
This function uses MetaSTAARliteWorker to generate variant summary statistics
and sparse LD matrices for gene-centric coding analysis.
}
