% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaSTAARlite_worker_cov.R
\name{MetaSTAARlite_worker_cov}
\alias{MetaSTAARlite_worker_cov}
\title{Generating sparse weighted covariance file using MetaSTAARlite (the "worker" step)}
\usage{
MetaSTAARlite_worker_cov(
  genotype,
  obj_nullmodel,
  cov_maf_cutoff = 0.05,
  qc_label = NULL,
  signif.digits = 3
)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants. If the input genotype matrix
is sparse (e.g. \code{dgCMatrix} format), it is assumed that it has been flipped to represent
minor allele coding.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function
in the \code{STAARpipeline} package, or the output from \code{fitNullModel} function in the \code{GENESIS} package
and transformed using the \code{\link{genesis2staar_nullmodel}} function in the \code{STAARpipeline} package.}

\item{cov_maf_cutoff}{a numeric value indicating the maximum minor allele frequency cutoff
under which the sparse weighted covariance file between variants is stored (default = 0.05).}

\item{qc_label}{a vector of quality control status for each variant in \code{genotype}, where a PASS variant
is labeled as "PASS". If \code{qc_label} is NULL, it is assumed that all variants are PASS variants in the study (default = NULL).}

\item{signif.digits}{an integer indicating the number of significant digits to be used
for storing the sparse weighted covariance file. If \code{signif.digits} is NULL,
it is assumed that no rounding will be performed (default = 3).}
}
\value{
\code{GTSinvG_rare}: the sparse matrix of all variants in the variant-set
whose minor allele frequency is below \code{cov_maf_cutoff} (the sparse weighted
covariance file).
}
\description{
The \code{MetaSTAARlite_worker_cov} function takes in genotype and the object
from fitting the null model to generate the sparse weighted
covariance file for the given variant-set.
}
\references{
Li, X., et al. (2023). Powerful, scalable and resource-efficient
meta-analysis of rare variant associations in large whole genome sequencing studies.
\emph{Nature Genetics}, \emph{55}(1), 154-164.
(\href{https://doi.org/10.1038/s41588-022-01225-6}{pub})
}
