% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaSTAARlite_worker_sumstat.R
\name{MetaSTAARlite_worker_sumstat}
\alias{MetaSTAARlite_worker_sumstat}
\title{Generating variant-level summary statistics file using MetaSTAARlite (the "worker" step)}
\usage{
MetaSTAARlite_worker_sumstat(
  genotype,
  ALT_AF = NULL,
  obj_nullmodel,
  variant_info,
  qc_label = NULL,
  annotation_phred = NULL,
  for_individual_analysis = FALSE
)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants. If the input genotype matrix
is sparse (e.g. \code{dgCMatrix} format), it is assumed that it has been flipped to represent
minor allele coding.}

\item{ALT_AF}{a numeric vector of alternate allele frequencies for each variant in \code{genotype}. 
This is required when \code{genotype} is in sparse format (e.g., \code{dgCMatrix}). 
This vector is the \code{ALT_AF} column from the \code{results_info} data frame returned by the \code{Genotype_flip_sp_extraction} function (default = NULL).}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function
in the \code{\link{STAARpipeline}} package, or the output from \code{fitNullModel} function in the \code{GENESIS} package
and transformed using the \code{\link{genesis2staar_nullmodel}} function in the \code{\link{STAARpipeline}} package.}

\item{variant_info}{a data frame or matrix of variant information (unique identifier)
with p rows (listed in the same order as the columns of \code{genotype}) and should contain
the following 4 columns: chromosome (chr), position (pos), reference allele (ref), and alternative allele (alt).}

\item{qc_label}{a vector of quality control status for each variant in \code{variant_info}, where a PASS variant
is labeled as "PASS". If \code{qc_label} is NULL, it is assumed that all variants are PASS variants in the study (default = NULL).}

\item{annotation_phred}{a data frame or matrix of functional annotation data
of dimension p*q (or a vector of a single annotation score with length p).
Continuous scores should be given in PHRED score scale, where the PHRED score
of j-th variant is defined to be -10*log10(rank(-score_j)/total) across the genome. (Binary)
categorical scores should be taking values 0 or 1, where 1 is functional and 0 is
non-functional. If not provided, STAAR will perform the
SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1), ACAT-V(1,25), ACAT-V(1,1)
and ACAT-O tests (default = NULL).}

\item{for_individual_analysis}{a logical value indicating whether it is used for individual (single-variant) meta-analysis
(default = FALSE).}
}
\value{
\code{sumstat}: the data frame of all variants in the variant-set or the list of individual variants
(the variant-level summary statistics file), including the following information: chromosome (chr), position (pos), reference allele (ref),
alternative allele (alt), quality control status (qc_label, optional), alternative allele count (alt_AC), minor allele count (MAC),
minor allele frequency (MAF), study sample size (N), score statistic (U), variance (V), variant annotations provided in
\code{annotation_phred}, and the low-rank decomposed component of the covariance file.
}
\description{
The \code{MetaSTAARlite_worker_sumstat} function takes in genotype, the object
from fitting the null model, and variant information (unique identifier)
to generate the variant-level summary statistics file for the given variant-set.
}
\references{
Li, X., et al. (2023). Powerful, scalable and resource-efficient
meta-analysis of rare variant associations in large whole genome sequencing studies.
\emph{Nature Genetics}, \emph{55}(1), 154-164.
(\href{https://doi.org/10.1038/s41588-022-01225-6}{pub})
}
