% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncRNA_MetaSTAARlite.R
\name{ncRNA_MetaSTAARlite}
\alias{ncRNA_MetaSTAARlite}
\title{Performs meta-analysis of long noncoding RNA (ncRNA) category using MetaSTAARlite}
\usage{
ncRNA_MetaSTAARlite(
  chr,
  gene_name,
  sample.sizes,
  ncRNA_sumstat_gene_list,
  ncRNA_cov_gene_list,
  cov_maf_cutoff,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  check_qc_label = FALSE,
  variant_type = c("SNV", "Indel", "variant"),
  Use_annotation_weights = TRUE,
  Annotation_name = NULL,
  silent = FALSE
)
}
\arguments{
\item{chr}{an integer which specifies the chromosome number.}

\item{gene_name}{a character which specifies the name of the ncRNA gene to be meta-analyzed using MetaSTAARlite.}

\item{sample.sizes}{a numeric vector with the length of \code{study.names}
indicating the sample size of each study.}

\item{ncRNA_sumstat_gene_list}{a list containing study-specific summary statistics corresponding to the specified gene.}

\item{ncRNA_cov_gene_list}{a list containing study-specific sparse weighted covariance matrices corresponding to the specified gene.}

\item{cov_maf_cutoff}{a numeric vector with the length of \code{study.names}
indicating the maximum minor allele frequency cutoffs under which the sparse weighted
covariance files between variants are stored.}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of meta-analyzing
a given variant-set (default = 2).}

\item{check_qc_label}{a logical value indicating whether variants need to be dropped according to \code{qc_label}
specified in \code{\link{ncRNA_MetaSTAARlite_worker}} (default = FALSE).}

\item{variant_type}{a character value specifying the type of variant included in the analysis. Choices include
"SNV", "Indel", or "variant" (default = "SNV").}

\item{Use_annotation_weights}{a logical value which determines if annotations will be used as weights or not (default = TRUE).}

\item{Annotation_name}{a character vector of annotation names used in MetaSTAARlite (default = NULL).}

\item{silent}{a logical value which determines if the report of error messages will be suppressed (default = FALSE).}
}
\value{
a list of data frames containing the MetaSTAAR p-values (including MetaSTAAR-O) corresponding to
the exonic and splicing category of the given ncRNA gene.
}
\description{
This function performs meta-analysis to detect associations between a
quantitative/dichotomous phenotype and the exonic and splicing category of an ncRNA gene by using the MetaSTAARlite pipeline.
For each coding functional category, the MetaSTAAR-O p-value is a p-value from an omnibus test
that aggregated SKAT-MS(1,25), SKAT-MS(1,1), Burden-MS(1,25), Burden-MS(1,1), ACAT-V-MS(1,25),
and ACAT-V-MS(1,1) together with p-values of each test weighted by each annotation
using Cauchy method.
}
\references{
Li, X., et al. (2023). Powerful, scalable and resource-efficient
meta-analysis of rare variant associations in large whole genome sequencing studies.
\emph{Nature Genetics}, \emph{55}(1), 154-164.
(\href{https://doi.org/10.1038/s41588-022-01225-6}{pub})

Li, Z., Li, X., et al. (2022). A framework for detecting noncoding
rare-variant associations of large-scale whole-genome sequencing studies.
\emph{Nature Methods}.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})

Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})
}
