% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_MetaSTAARlite_worker.R
\name{coding_MetaSTAARlite_worker}
\alias{coding_MetaSTAARlite_worker}
\title{Generates summary statistics of coding functional categories using MetaSTAARlite}
\usage{
coding_MetaSTAARlite_worker(
  chr,
  gene_name,
  genofile,
  obj_nullmodel,
  genes,
  known_loci = NULL,
  cov_maf_cutoff = 0.05,
  signif.digits = NULL,
  QC_label = "annotation/filter",
  check_qc_label = TRUE,
  variant_type = c("SNV", "Indel", "variant"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = TRUE,
  Annotation_name = NULL,
  silent = FALSE
)
}
\arguments{
\item{chr}{an integer which specifies the chromosome number.}

\item{gene_name}{a character which specifies the name of the gene to be meta-analyzed using MetaSTAARlite.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function
in the \code{STAARpipeline} package, or the output from \code{fitNullModel} function in the \code{GENESIS} package
and transformed using the \code{\link{genesis2staar_nullmodel}} function in the \code{STAARpipeline} package.}

\item{genes}{the \code{genes_info} object from the \code{STAARpipeline} package.}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis. Should
contain four columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT) (default = NULL).}

\item{cov_maf_cutoff}{a numeric value indicating the maximum minor allele frequency cutoff
under which the sparse weighted covariance file between variants is stored (default = 0.05).}

\item{signif.digits}{an integer indicating the number of significant digits to be used
for storing the sparse weighted covariance file. If \code{signif.digits} is NULL,
it is assumed that no rounding will be performed (default = NULL).}

\item{QC_label}{a character specifying the channel name of the QC label in the GDS/aGDS file
(default = "annotation/filter").}

\item{check_qc_label}{a logical value indicating whether variants need to be dropped according to \code{qc_label}.
If \code{check_qc_label} is FALSE, then the summary statistics will be stored for PASS variants from the study.
If \code{check_qc_label} is TRUE, then the summary statistics will be stored for all variants from the study,
together will an additional column of \code{qc_label} (default = TRUE).}

\item{variant_type}{a character value specifying the type of variant included in the analysis. Choices include
"SNV", "Indel", or "variant" (default = "SNV").}

\item{Annotation_dir}{a character specifying the channel name of the annotations in the aGDS file
(default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the annotation name and the corresponding
channel name in the aGDS file.}

\item{Use_annotation_weights}{a logical value which specifies if annotations will be used as weights
or not (default = TRUE).}

\item{Annotation_name}{a character vector of annotation names used in MetaSTAARlite (default = NULL).}

\item{silent}{a logical value which determines if the report of error messages will be suppressed (default = FALSE).}
}
\value{
a list of the following objects corresponding to each coding functional category of the given gene:
(1) the data frame of all variants in the variant-set (the variant-level summary statistics file),
including the following information: chromosome (chr), position (pos), reference allele (ref),
alternative allele (alt), quality control status (qc_label, optional), alternative allele count (alt_AC), minor allele count (MAC),
minor allele frequency (MAF), study sample size (N), score statistic (U), variance (V), variant annotations specified in
\code{Annotation_name}, and the low-rank decomposed component of the covariance file;
(2) the sparse matrix of all variants in the variant-set whose minor allele frequency is below \code{cov_maf_cutoff} (the sparse weighted
covariance file); (3) the summary statistics and covariance matrices corresponding to the specified gene for variants to be conditioned on
in \code{known_loci}.
}
\description{
This function uses MetaSTAARlite to generate variant-level summary statistics
and sparse covariance matrices for coding functional categories of a gene.
}
\references{
Li, X., et al. (2023). Powerful, scalable and resource-efficient
meta-analysis of rare variant associations in large whole genome sequencing studies.
\emph{Nature Genetics}, \emph{55}(1), 154-164.
(\href{https://doi.org/10.1038/s41588-022-01225-6}{pub})

Li, Z., Li, X., et al. (2022). A framework for detecting noncoding
rare-variant associations of large-scale whole-genome sequencing studies.
\emph{Nature Methods}.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})

Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})
}
