% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Centric_Coding_Results_Summary_meta.R
\name{Gene_Centric_Coding_Results_Summary_meta}
\alias{Gene_Centric_Coding_Results_Summary_meta}
\title{Generates summary table and/or visualization for the meta-analysis of coding functional categories that was conducted.}
\usage{
Gene_Centric_Coding_Results_Summary_meta(
  gene_centric_coding_jobs_num,
  input_path,
  output_path,
  gene_centric_results_name,
  alpha = 2.5e-06,
  manhattan_plot = FALSE,
  QQ_plot = FALSE
)
}
\arguments{
\item{gene_centric_coding_jobs_num}{an integer which specifies the number of jobs done in the meta-analysis.}

\item{input_path}{a character which specifies the file path to the meta-analysis results files.}

\item{output_path}{a character which specifies the file path to the desired location of the produced summary table and visualizations.}

\item{gene_centric_results_name}{a character which specifies the name (excluding the jobs number) of the meta-analysis results files.}

\item{alpha}{a numeric value which specifies the desired significance threshold. Default is 2.5E-06.}

\item{manhattan_plot}{a logical value which determines if a Manhattan plot is generated. Default is FALSE.}

\item{QQ_plot}{a logical value which determines if a QQ plot is generated. Default is FALSE.}
}
\description{
This function generates a summary table, Manhattan plot, and QQ plot for the meta-analysis of coding functional categories that was conducted
based on the parameters provided by the user.
}
