% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_analysis_MetaSTAARlite_cond.R
\name{individual_analysis_MetaSTAARlite_cond}
\alias{individual_analysis_MetaSTAARlite_cond}
\title{Performs conditional meta-analysis of individual variants using MetaSTAARlite}
\usage{
individual_analysis_MetaSTAARlite_cond(
  individual_results,
  sample.sizes,
  sumstat.list,
  covcond.list,
  mac_cutoff = 20,
  effect.cond = c("homogeneous", "heterogeneous"),
  check_qc_label = TRUE
)
}
\arguments{
\item{individual_results}{a dataframe containing the (significant) results of the individual variant meta-analysis.}

\item{sample.sizes}{a numeric vector with the length of \code{study.names}
indicating the sample size of each study.}

\item{sumstat.list}{a list containing study-specific summary statistics from all participating studies.}

\item{covcond.list}{a list containing study-specific summary statistics and covariance matrices
for variants to be conditioned on from all participating studies.}

\item{mac_cutoff}{an integer specifying the cutoff of minimum combined minor allele count in
defining individual variants (default = 20).}

\item{effect.cond}{a character value indicating the effects of variants to be adjusted for
in conditional analysis are "homogeneous" or "heterogeneous" (default = "homogeneous").}

\item{check_qc_label}{a logical value indicating whether variants need to be dropped according to \code{qc_label}
specified in \code{\link{individual_analysis_MetaSTAARlite_worker}} (default = TRUE).}
}
\value{
a data frame containing the conditional meta-analysis score test p-value and the estimated effect size
of the alternative allele for each (significant) individual variant in \code{individual_results}.
}
\description{
This function performs conditional meta-analysis to detect associations between a quantitative/dichotomous phenotype
and each (significant) individual variant by using conditional score test.
}
