% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_analysis_MetaSTAARlite_worker.R
\name{individual_analysis_MetaSTAARlite_worker}
\alias{individual_analysis_MetaSTAARlite_worker}
\title{Generates summary statistics of individual variants using MetaSTAARlite}
\usage{
individual_analysis_MetaSTAARlite_worker(
  chr,
  start_loc,
  end_loc,
  genofile,
  obj_nullmodel,
  known_loci = NULL,
  subsegment.size = 50000,
  QC_label = "annotation/filter",
  check_qc_label = FALSE,
  variant_type = c("variant", "SNV", "Indel"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = c(FALSE, TRUE),
  Annotation_name = NULL,
  silent = FALSE
)
}
\arguments{
\item{chr}{an integer which specifies the chromosome number.}

\item{start_loc}{an integer which specifies the starting location of variants to be analyzed.}

\item{end_loc}{an integer which specifies the end location of variants to be analyzed.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function
in the \code{\link{STAARpipeline}} package, or the output from \code{fitNullModel} function in the \code{GENESIS} package
and transformed using the \code{\link{genesis2staar_nullmodel}} function in the \code{\link{STAARpipeline}} package.}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis. Should
contain four columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT) (default = NULL).}

\item{subsegment.size}{a numeric value which specifies the size of each subsegment for computation (default = 5e4).}

\item{QC_label}{a character specifying the channel name of the QC label in the GDS/aGDS file
(default = "annotation/filter").}

\item{check_qc_label}{a logical value indicating whether variants need to be dropped according to \code{qc_label}.
If \code{check_qc_label} is FALSE, then the summary statistics will be stored for PASS variants from the study.
If \code{check_qc_label} is FALSE, then the summary statistics will be stored for all variants from the study,
together will an additional column of \code{qc_label} (default = FALSE).}

\item{variant_type}{a character value specifying the type of variant included in the analysis. Choices include
"SNV", "Indel", or "variant" (default = "variant").}

\item{Annotation_dir}{a character specifying the channel name of the annotations in the aGDS file
(default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the annotation name and the corresponding
channel name in the aGDS file.}

\item{Use_annotation_weights}{a logical value which specifies if annotations will be used as weights
or not (default = FALSE).}

\item{Annotation_name}{a character vector of annotation names used in MetaSTAARlite (default = NULL).}

\item{silent}{a logical value which determines if the report of error messages will be suppressed (default = FALSE).}
}
\value{
a list of the following objects corresponding to each individual variant in the given genetic region:
(1) the data frame of all variants in the variant-set (the variant-level summary statistics file),
including the following information: chromosome (chr), position (pos), reference allele (ref),
alternative allele (alt), quality control status (qc_label, optional), alternative allele count (alt_AC), minor allele count (MAC),
minor allele frequency (MAF), study sample size (N), score statistic (U), variance (V), variant annotations specified in
\code{Annotation_name} (optional), and the low-rank decomposed component of the covariance file;
(2) the summary statistics and covariance matrices corresponding to the specified gene for variants to be conditioned on
in \code{known_loci}.
}
\description{
This function uses MetaSTAARlite to generate variant-level summary statistics
for individual variants of interest.
}
\references{
Li, X., et al. (2023). Powerful, scalable and resource-efficient
meta-analysis of rare variant associations in large whole genome sequencing studies.
\emph{Nature Genetics}, \emph{55}(1), 154-164.
(\href{https://doi.org/10.1038/s41588-022-01225-6}{pub})

Li, Z., Li, X., et al. (2022). A framework for detecting noncoding
rare-variant associations of large-scale whole-genome sequencing studies.
\emph{Nature Methods}.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})
}
